Created new files data.js, util.js Changed code in main.js
